Maven - Intellij - Antlr4 Integration
Antlr4 and project code compiles with Maven build tools, scripts are included to do this manually or the Intellij projects files are included so it can be imported with minimal work. Eclipse should work with Maven exporting, as well.

Issues still up: 

Report is not formatted right as of now.

GUI is not in this code yet.

Documentation is not yet complete (may be needed for after main presentation).

And more...